requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.9.1
click==8.1.7
coverage==7.3.2
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.10.0
gunicorn==21.2.0
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
uritemplate==4.1.1
